Implementing S3 connection

+++ gjblog/src/main/java/com/zipcode/gjblog/repository/BlogRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlogRepository extends CrudRepository<Post,Integer> {
+public interface BlogRepository extends CrudRepository<Post,Long> {
 
 }
